+++ Samples/iOS-Swift/UITests/Sources/UserFeedbackUITests.swift
@@ -23,17 +23,32 @@ class UserFeedbackUITests: BaseUITest {
             SentrySDKOverrides.Special.disableEverything.rawValue,
             
             // write base64-encoded data into the envelope file for attachments instead of raw bytes, specifically for images. this way the entire envelope contents can be more easily passed as a string through the text field in the app to this process for validation.
-            SentrySDKOverrides.Other.base64AttachmentData.rawValue
+            SentrySDKOverrides.Attachments.base64Data.rawValue
         ])
         continueAfterFailure = true
     }
 }
 
 extension UserFeedbackUITests {
+    private func launchAppWithDeprecatedWidget(args: [String] = [], env: [String: String] = [:]) {
+        var launchArguments = args
+        if !launchArguments.contains(SentrySDKOverrides.Feedback.disableAutoInject.rawValue) {
+            launchArguments.append(SentrySDKOverrides.Feedback.disableAutoInject.rawValue)
+        }
+        launchApp(args: launchArguments, env: env)
+        showDeprecatedWidget()
+    }
+
+    private func showDeprecatedWidget() {
+        extrasAreaTabBarButton.tap()
+        app.buttons["io.sentry.ui-test.button.show-widget"].tap()
+        XCTAssertTrue(widgetButton.waitForExistence(timeout: 1))
+    }
+
     // MARK: Tests ensuring correct appearance
     
     func testUIElementsWithDefaults() {
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue])
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue])
         // widget button text
         XCTAssertTrue(app.otherElements["Report a Bug"].exists)
         
@@ -60,7 +75,7 @@ extension UserFeedbackUITests {
     }
     
     func testUIElementsWithCustomizations() {
-        launchApp()
+        launchAppWithDeprecatedWidget()
         
         // widget button text
         XCTAssertTrue(app.otherElements["Report Jank"].exists)
@@ -88,11 +103,11 @@ extension UserFeedbackUITests {
     }
     
     func testPrefilledUserInformation() throws {
-        launchApp(args: [
+        launchAppWithDeprecatedWidget(args: [
             SentrySDKOverrides.Feedback.allDefaults.rawValue
         ], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: "ui test user",
-            SentrySDKOverrides.Other.userEmail.rawValue: "ui-testing@sentry.io"
+            SentrySDKOverrides.Scope.userFullName.rawValue: "ui test user",
+            SentrySDKOverrides.Scope.userEmail.rawValue: "ui-testing@sentry.io"
         ])
         
         widgetButton.tap()
@@ -101,11 +116,11 @@ extension UserFeedbackUITests {
     }
     
     func testNoPrefilledUserInformation() throws {
-        launchApp(args: [
+        launchAppWithDeprecatedWidget(args: [
             SentrySDKOverrides.Feedback.noUserInjection.rawValue
         ], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: "ui test user",
-            SentrySDKOverrides.Other.userEmail.rawValue: "ui-testing@sentry.io"
+            SentrySDKOverrides.Scope.userFullName.rawValue: "ui test user",
+            SentrySDKOverrides.Scope.userEmail.rawValue: "ui-testing@sentry.io"
         ])
         
         widgetButton.tap()
@@ -118,7 +133,7 @@ extension UserFeedbackUITests {
     // MARK: Tests validating happy path / successful submission
 
     func testSubmitFullyFilledCustomForm() throws {
-        launchApp(args: [
+        launchAppWithDeprecatedWidget(args: [
             SentrySDKOverrides.Feedback.noUserInjection.rawValue
         ])
 
@@ -158,9 +173,9 @@ extension UserFeedbackUITests {
         let testName = "Andrew"
         let testContactEmail = "andrew.mcknight@sentry.io"
         
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: testName,
-            SentrySDKOverrides.Other.userEmail.rawValue: testContactEmail
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
+            SentrySDKOverrides.Scope.userFullName.rawValue: testName,
+            SentrySDKOverrides.Scope.userEmail.rawValue: testContactEmail
         ])
 
         try retrieveAppUnderTestApplicationSupportDirectory()
@@ -246,9 +261,9 @@ extension UserFeedbackUITests {
         let testName = "Andrew"
         let testContactEmail = "andrew.mcknight@sentry.io"
         
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: testName,
-            SentrySDKOverrides.Other.userEmail.rawValue: testContactEmail
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
+            SentrySDKOverrides.Scope.userFullName.rawValue: testName,
+            SentrySDKOverrides.Scope.userEmail.rawValue: testContactEmail
         ])
 
         try retrieveAppUnderTestApplicationSupportDirectory()
@@ -275,9 +290,9 @@ extension UserFeedbackUITests {
         let testName = "Andrew"
         let testContactEmail = "andrew.mcknight@sentry.io"
         
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: testName,
-            SentrySDKOverrides.Other.userEmail.rawValue: testContactEmail
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
+            SentrySDKOverrides.Scope.userFullName.rawValue: testName,
+            SentrySDKOverrides.Scope.userEmail.rawValue: testContactEmail
         ])
 
         try retrieveAppUnderTestApplicationSupportDirectory()
@@ -296,6 +311,7 @@ extension UserFeedbackUITests {
         try assertOnlyHookMarkersExist(names: [.onFormClose])
         
         // displaying the form again ensures the widget button still works afterwards; also assert that the fields are in their default state to ensure the entered data is not persisted between displays
+        XCTAssertTrue(widgetButton.waitForExistence(timeout: 1))
         widgetButton.tap()
         
         XCTAssertEqual(try XCTUnwrap(nameField.value as? String), testName)
@@ -312,9 +328,9 @@ extension UserFeedbackUITests {
         let testName = "Andrew"
         let testContactEmail = "andrew.mcknight@sentry.io"
         
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: testName,
-            SentrySDKOverrides.Other.userEmail.rawValue: testContactEmail
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
+            SentrySDKOverrides.Scope.userFullName.rawValue: testName,
+            SentrySDKOverrides.Scope.userEmail.rawValue: testContactEmail
         ])
 
         try retrieveAppUnderTestApplicationSupportDirectory()
@@ -352,7 +368,7 @@ extension UserFeedbackUITests {
         let testMessage = "UITest user feedback"
         fillInFields(testMessage)
         
-        submit()
+        submit(waitingForWidget: false)
         
         try assertEnvelopeContents(testMessage, attachments: true)
     }
@@ -368,7 +384,7 @@ extension UserFeedbackUITests {
         let testMessage = "UITest user feedback"
         fillInFields(testMessage)
         
-        submit()
+        submit(waitingForWidget: false)
         
         try assertEnvelopeContents(testMessage)
     }
@@ -376,7 +392,7 @@ extension UserFeedbackUITests {
     // MARK: Tests validating error cases
     
     func testSubmitWithNoFieldsFilledDefault() throws {
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue])
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue])
 
         try retrieveAppUnderTestApplicationSupportDirectory()
         try assertHookMarkersNotExist()
@@ -396,7 +412,7 @@ extension UserFeedbackUITests {
     }
     
     func testSubmitWithNoFieldsFilledEmailAndMessageRequired() throws {
-        launchApp(args: [
+        launchAppWithDeprecatedWidget(args: [
             SentrySDKOverrides.Feedback.requireEmail.rawValue,
             SentrySDKOverrides.Feedback.noUserInjection.rawValue
         ])
@@ -423,7 +439,7 @@ extension UserFeedbackUITests {
     }
     
     func testSubmitWithNoFieldsFilledAllRequired() throws {
-        launchApp(args: [
+        launchAppWithDeprecatedWidget(args: [
             SentrySDKOverrides.Feedback.requireEmail.rawValue,
             SentrySDKOverrides.Feedback.requireName.rawValue,
             SentrySDKOverrides.Feedback.noUserInjection.rawValue
@@ -450,7 +466,7 @@ extension UserFeedbackUITests {
     }
     
     func testSubmitOnlyWithOptionalFieldsFilled() throws {
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue])
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue])
 
         try retrieveAppUnderTestApplicationSupportDirectory()
         try assertHookMarkersNotExist()
@@ -472,9 +488,9 @@ extension UserFeedbackUITests {
         let testName = "Andrew"
         let testContactEmail = "andrew.mcknight@sentry.io"
         
-        launchApp(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
-            SentrySDKOverrides.Other.userFullName.rawValue: testName,
-            SentrySDKOverrides.Other.userEmail.rawValue: testContactEmail
+        launchAppWithDeprecatedWidget(args: [SentrySDKOverrides.Feedback.allDefaults.rawValue], env: [
+            SentrySDKOverrides.Scope.userFullName.rawValue: testName,
+            SentrySDKOverrides.Scope.userEmail.rawValue: testContactEmail
         ])
         
         try retrieveAppUnderTestApplicationSupportDirectory()
@@ -591,13 +607,15 @@ extension UserFeedbackUITests {
 
 // MARK: Form hook test helpers
 extension UserFeedbackUITests {
-    func submit(expectingError: Bool = false, usingCustomButton: Bool = false) {
+    func submit(expectingError: Bool = false, usingCustomButton: Bool = false, waitingForWidget: Bool = true) {
         sendButton.tap()
         if !expectingError {
             if usingCustomButton {
                 customButton.waitForExistence("Form should have been dismissed and custom button should be visible again.")
-            } else {
+            } else if waitingForWidget {
                 widgetButton.waitForExistence("Form should have been dismissed and widget button should be visible again.")
+            } else {
+                XCTAssertFalse(sendButton.waitForExistence(timeout: 1))
             }
         }
     }