+++ Samples/iOS-SwiftUI/App/Sources/ContentView.swift
@@ -223,6 +223,10 @@ struct ContentView: View {
                             NavigationLink(destination: FormScreen()) {
                                 Text("Form Screen")
                             }
+
+                            NavigationLink(destination: FeedbackScreen()) {
+                                Text("Feedback Screen")
+                            }
                         }
                         .background(Color.white)
                     }